move the mapper to the util package.
