+++ src/main/kotlin/data/remote/mongo/AuthMongoImpl.kt
@@ -5,7 +5,7 @@ import org.bson.Document
 import org.example.data.AuthProvider
 import org.example.entity.UserEntity
 import org.example.entity.UserType
-import org.example.utils.MongoExceptionHandler
+import org.example.data.util.exception.MongoExceptionHandler
 import org.example.utils.PlanMateException
 import java.util.*
