+++ src/main/kotlin/data/remote/mongo/TaskMongoDBImpl.kt
@@ -7,7 +7,7 @@ import kotlinx.datetime.toLocalDateTime
 import org.bson.Document
 import org.example.data.RemoteDataSource
 import org.example.entity.TaskEntity
-import org.example.utils.MongoExceptionHandler
+import org.example.data.util.exception.MongoExceptionHandler
 import org.example.utils.PlanMateException
 import java.util.*
